Adds bestSkaters query and PlayerCardMeta interface
